Move propagators into their own files
